Made name optional for both garments and outfits, and pulled inline dto's out for outfit.
