+++ src/wardrobe/dto/create-outfit.dto.ts
@@ -1,0 +1,7 @@
+import { OutfitSlot } from 'src/dal/entity/outfit.entity';
+
+export interface CreateOutfitDto {
+  name?: string;
+  notes?: string;
+  slots?: OutfitSlot[];
+}
