+++ src/wardrobe/dto/update-outfit.dto.ts
@@ -1,0 +1,7 @@
+import { OutfitSlot } from 'src/dal/entity/outfit.entity';
+
+export interface UpdateOutfitDto {
+  name?: string;
+  notes?: string;
+  slots?: OutfitSlot[];
+}
