+++ views/outfits/form.hbs
@@ -82,7 +82,6 @@
         class="input input-bordered w-full"
         value="{{#if outfit}}{{outfit.name}}{{/if}}"
         placeholder="e.g. Summer Date Night"
-        required
       />
     </div>
